Game Over provisional

+++ Gloton Galactico/Game/Source/Player.h
@@ -38,7 +38,7 @@ public:
 	
 	bool godMode;
 
-	float vidas;
+	int vidas;
 
 private:
